server/requirements.txt
astroid==1.4.8
Django==1.10
django-redis-cache==1.7.0
djangorestframework==3.4.6
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.2
pkg-resources==0.0.0
psycopg2==2.6.2
pylint==1.6.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
redis==2.10.5
six==1.10.0
uWSGI==2.0.13.1
wrapt==1.10.8
